Fix a bug in H3SVGWriter::writeData(const Reward&).
The type of the reward was incorrectly serialized as 1 byte, but in H3SVG
it is serialized as 4 bytes.

+++ src/h3mtxt/H3Writer/H3SVGWriter/Reward.cpp
@@ -80,7 +80,7 @@ namespace h3svg
 
   void H3SVGWriter::writeData(const Reward& reward) const
   {
-    writeData(reward.type());
+    writeData(static_cast<std::uint32_t>(reward.type()));
     std::visit([this] <RewardType T> (const RewardDetails<T>& details)
                { writeData(details); },
                reward.details);
